requirements.txt
fastapi
uvicorn[standard]
python-multipart
google-auth
google-auth-oauthlib
google-api-python-client
google-analytics-admin
google-analytics-data
requests
python-dotenv
itsdangerous
SQLAlchemy
jinja2
pydantic-settings